+++ src/test/kotlin/study/junit/PersonTest.kt
@@ -1,8 +1,9 @@
-package step1
+package study.junit
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
+import study.Person
 
 class PersonTest {
     @Test
